pyproject.toml
[project]
name = "multi-lingual-qac"
version = "0.1.0"
description = "Multi-lingual chemical Question-Answer-Context pipeline"
readme = "docs/MULTILINGUAL_CHEMICAL_QAC_PLAN.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=2.14.0",
    "google-cloud-bigquery>=3.0.0",
    "mteb>=2.12.15",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
]

[project.scripts]
qac = "main:main"
